module support, basic tracepoint support
Refactor the entire plumbing around resolving/validating/compiling
calls (functions) and methods. All functions now belong to a module and
it is up to the provider to include the relevant modules to its scope.

A fully qualified call, i.e. `module.func()`, is only valid if the
module in question is able to resolve `func`. "Bare" calls,
i.e. `func()`, are still allowed. In this case `func` will resolve to
the implementation of the first module in the provider's module list
that is able to resolve it.

The reason for this is that we want to export a tracepoint's fields as
regular functions in `ply`. Something like this:

    trace:sched/sched_switch { $s[prev_comm].count() }

The problem is that the field names are not known at compile time. There
is thus no way for `ply` to check that the global namespace is free of
name conflicts. Indeed, some tracepoints have fields called `pid`, which
would conflict with `ply`s built in version.

With the module support the tracepoint's `pid` will shadow the built in,
but it is still available using `common.pid`.

`ply` is now able to attach to tracepoints, but it is not possible to
extract their fields yet. That is next on the todo.

PS. The author is aware that this mega-commit is not exactly "clean". If
given the chance, he might have tried to get a few more intermediate
commits in. Ah well... DS.