unroll: add unroll statement
The kernel does not allow backward branches, making the generic
`for`-loop impossible to implement. Work around it by having an `unroll`
statement which just copy+pastes the containing block of code `N` times.

I.e. the following code:

    kprobe:SyS_exec* {
        unroll(2) {
	    printf("double trouble\n")
	}
    }

Would print the string twice on each `exec`. But this is generated at
compile time, thus the the unroll count must be an integer constant.