evpipe: generic event pipe
Due to an oversight by yours truly, `printf()` was implemented in quite
a hacky way using a regular array map. The proper way to do this is to
use `perf_event_output()`. Correct this mistake. While we're here
prepare for other types of events beside `printf()`. `exit()` comes to
mind and I'm sure there are others.