lang: add support for "default probes"
Allow each provider to specify a default probe if none is supplied by
the user. I.e. the user can simply say:

    kprobe:SyS_open

And get a probe that prints out a message when the probe is hit. This is
very useful for quickly testing if a certain function is ever called.