perf-probe: remove dead code, use group_fd
Perf events are never per-cpu for any of the current use-cases, so
remove the dead code for dealing with it.

Also, tie all probes in a wildcard match to the same group_fd so that
they are scheduled together.

TODO: Use the same group_fd for all probes in a script.