parens should not be required for probe predicates
a probe predicate is defined as an expr in lex.l; however
exprs can contain '/' division symbols, which are also used
to bracket a predicate.  This causes a shift-reduce conflict
for flex, but we can resolve it by having a '/' followed by
a '{' resolve to a CLOSEPRED token while other instances of '/'
resolve to a DIV token.  This change allows us to define probe
predicates without having to use parentheses, e.g. in
syscall-fail.ply.

Signed-off-by: Alan Maguire <alan.maguire@oracle.com>