ir: add perf_event_output helper, evpipe: delay creation
The evpipe was created too early, making it impossible to do dry-runs
as UID!=0. Also, create a dummy symbol for it so we can use
ir_emit_ldmap to get to the mapfd.

Create a helper to generate perf events. This will be used to send all
kinds of events to the ply process such as exit requests, printfs etc.