string: add virtual attribute
Some strings are in the AST but are not to be encoded in the resulting
BPF program. Traditionally this was limited to struct member names, so
we explicitly checked for that. We want the same behavior for format
strings to the upcoming printf. Thus make it a generic attribute on
the node so we don't have to keep updating the string loader with new
conditions.