printxf: handle va_lists in a portable manner
va_list can be a pointer, an array, or a short garden gnome. So we
can't make any assumptions about it behaves. On x86_64 a va_list will
be a reference, so passing it to vfprintf will update the callers
version. On ARM this is not the case. So always send a copy and then
parse enough of the specifier to figure out what lies to tell va_arg
to make the local va_list match the state that vfprintf will have left
its copy.