xprobe: avoid untracable symbols on wildcard matches
No function in the init-section of the kernel will ever be tracable,
so skip them.

Additionally, skip all symbols containing a dot (.), which is
typically generated internally by the compiler.

TODO: Find some way of actually finding out if the "notrace" attribute
is set on a symbol or not.

Partially fixes iovisor/ply#50.