stack: fix segfault on multiple references
Using the symbol's priv pointer as an indicator of whether we've
rewritten this node is incorrect, since the symbol is shared across
all references to it. Kind of the point.

Instead look at the args to see if ctx and the map have been
added. Then look at the symbol's type to see if this is the first
reference, if so setup the map and backtrace buffer.