kretprobe/retval: keep type information after ast rewrite
Previously an expression like `retval < 0` would generate an unsigned
comparison as `retval` was rewritten to `regs->rax` or similar, which
is unsigned (`reg_t`).

Instead keep the information by inserting the real expression under
the original node, and use the new `func_pass_ir_post` helper to pass
it upwards.