configure: map all arm* toolchain prefixes to arm
Particularly for ARM, toolchains can have a variety of names for the
CPU part of the triplet. "arm", "armhf", "armv7", "armv7l", "armv7a",
etc. etc. etc. This breaks the build as a direct mapping from CPU name
to architecture name in src/libply/arch/ is assumed.

Insert a mangling step in which arm* can all be mapped to arm. We can
extend this in the future for other architectures if necessary.