libply: convert providers to use a list
For static linking, it cannot use linker symbols as it's not
available.  And providers are not referenced from the main binary
directly, so the linker will omit the object.

Add a reference to each providers in the provider_init() and call it
in the ply_init() which is called in the main binary.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>