type: Allow non-standard struct layouts
Some tracepoint data uses a layout that is neither packed nor
"regular". The actual offset is available in the tracepoint's "format"
description, but ply has not been able to make use of it.

This is a first step towards changing that. Store the offset of each
field rather than computing it. This means we can just copy whatever
the format says. It also makes it easier to import types from BTF and
DWARF in the future.