isstring: Fix off-by-one
Previously, an array of size 4 with the following content:

74 65 73 74

Would be classified as a string ("test"). In fact, it is not as there
is no space to store the NUL. This would sometimes lead to weird
trailing output on truncated strings.

Ensure that the NUL byte also fits inside the string. Notably, this
also means that truncated strings will be output as hexdumps, which
makes it obvious that the truncation occurred.