ply: Fix SIGCHLD from inferior exit sporadically being lost
Intermittently, a command like `ply -c "sleep 1" '<some-probe>'` would
not terminate after the sleep process terminated.

This was because the old logic relied on SIGCHLD being delivered to
ply while blocked on ply_loop(). In other cases, it would be lost and
the ply process would hang until terminated with a SIGINT or similar.

Fix this by letting the user of libply get a copy of the fds to poll
and then calling ply_service() whenever an event comes in. This way,
we can fold in the signal monitoring into the main loop using a
signalfd, which libply does not have to be aware of.